requirements.txt
requests>=2.28
cryptography>=41.0
python-dotenv>=1.0

# backend
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-multipart>=0.0.9
httpx>=0.27

# LLM (use either; whichever key is set wins)
anthropic>=0.40
openai>=1.40
